requirements.txt
fastapi==0.85.1
uvicorn[standard]==0.19.0
aiohttp==3.8.3
aiofiles==22.1.0
python-multipart==0.0.5
PyYAML==6.0
